Complete display ad dropdown usage
