+++ app/views/articles/show.html.erb
@@ -195,12 +195,24 @@
       <% cache("article-bottom-content-#{rand(5)}-#{@article.id}-#{user_signed_in?}-#{(@organization || @user).latest_article_updated_at}", expires_in: 15.minutes) do %>
         <% @display_ad = DisplayAd.for_display("post_comments", user_signed_in?, @article.decorate.cached_tag_list_array) %>
         <% if @display_ad %>
-          <div class="crayons-card crayons-card--secondary crayons-sponsorship-article text-styles"
+          <div class="crayons-card crayons-card--secondary crayons-sponsorship-article"
                data-display-unit data-id="<%= @display_ad.id %>"
                data-category-click="<%= DisplayAdEvent::CATEGORY_CLICK %>"
                data-category-impression="<%= DisplayAdEvent::CATEGORY_IMPRESSION %>"
                data-context-type="<%= DisplayAdEvent::CONTEXT_TYPE_HOME %>">
-            <%= @display_ad.processed_html.html_safe %>
+            <div class="crayons-sponsorship__header relative">
+              <button id="sponsorship-dropdown-trigger-<%= @display_ad.id %>" aria-controls="sponsorship-dropdown-<%= @display_ad.id %>" aria-expanded="false" aria-haspopup="true" 
+                  class="dropBtn crayons-sponsorship__dropdown crayons-btn crayons-btn--s crayons-btn--ghost crayons-btn--icon"
+                  aria-label="Sponsorship dropdown">
+                  <%= crayons_icon_tag("overflow-vertical", class: "pointer-events-none") %>
+              </button>
+              <div id="sponsorship-dropdown-<%= @display_ad.id %>" class="crayons-dropdown right-0">
+                <%= render "shared/display_ad_dropdown_content" %>
+              </div>
+            </div>
+            <div class="p-4 pt-0 text-styles">
+              <%= @display_ad.processed_html.html_safe %>
+            </div>
           </div>
         <% end %>
       <% end %>
