+++ app/views/shared/_display_ad_dropdown_content.html.erb
@@ -4,7 +4,7 @@
         target="_blank"
         class="crayons-link crayons-link--block"
         rel="noopener"
-        href='/billboards'>
+        href='/billboard'>
         <%= crayons_icon_tag("info", aria_hidden: true, class: "c-btn__icon") %>
         <%= I18n.t("display_ads.what_is_a_billboard") %>
         </a>
